+++ src/kspdg/private_src/python3_9/pyarmor_runtime_000000/__init__.py
@@ -1,0 +1,2 @@
+# Pyarmor 8.5.11 (trial), 000000, 2024-08-09T11:36:40.809784
+from .pyarmor_runtime import __pyarmor__