+++ tests/ksp_ingame_tests/test_lbg1_i2.py
@@ -10,12 +10,23 @@ Test that require:
 """
 import pytest
 import time
+import sys
 import numpy as np
 
 import kspdg.utils.constants as C
 from kspdg.lbg1.lg0_envs import LBG1_LG0_I2_Env
 from kspdg.lbg1.lg1_envs import LBG1_LG1_I2_Env
 from kspdg.lbg1.lg2_envs import LBG1_LG2_I2_Env
+
+if sys.version_info[:2] == (3, 12):
+    # Python 3.12
+    from kspdg.private_src.python3_12.kspdg_envs.lbg1.lg3_envs import LBG1_LG3_I2_Env
+elif sys.version_info[:2] == (3, 9):
+    # Python 3.9
+    from kspdg.private_src.python3_9.kspdg_envs.lbg1.lg3_envs import LBG1_LG3_I2_Env
+else:
+    # Handle other versions or raise an error
+    raise ImportError(f"LBG1_LG3_I2_Env requires python 3.9 or 3.12, got {sys.version}")
 
 @pytest.fixture
 def lbg1_lg0_i2_env():
@@ -41,6 +52,14 @@ def lbg1_lg2_i2_env():
     yield env
     env.close()
 
+@pytest.fixture
+def lbg1_lg3_i2_env():
+    """setup and teardown of the LBG1_LG2_I2_Env object connected to kRPC server"""
+    env = LBG1_LG3_I2_Env()
+    env.reset()
+    yield env
+    env.close()
+
 def test_smoketest_lg0(lbg1_lg0_i2_env):
     """Ensure no errors are thrown from starting LG0 environment"""
     pass
@@ -51,6 +70,10 @@ def test_smoketest_lg1(lbg1_lg1_i2_env):
 
 def test_smoketest_lg2(lbg1_lg2_i2_env):
     """Ensure no errors are thrown from starting LG2 environment"""
+    pass
+
+def test_smoketest_lg3(lbg1_lg3_i2_env):
+    """Ensure no errors are thrown from starting LG3 environment"""
     pass
 
 def test_observation_0(lbg1_lg0_i2_env):