+++ CHANGELOG.md
@@ -13,10 +13,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `juliacall` dependency including basic unit tests to check proper installation
 - new optional dependency groups in `pyproject.toml`: `adv_bots` with juliacall dependency and `full` which installs all other dependency groups
 - `iLQGames.jl` dependency and python script for partially automating the install process
+- `private_src` directory to hold all compiled and/or obfuscated source code (e.g. to obfuscate bot policies in environments)
+- 'lg3_envs.py` that use ilqgames-based Guard agent
+- 'matplotlib' dependency for testing group
+
 
 ### Fixed
 
 ### Changed
+
+- Updated `example_eval_cfg.yaml` and related instructions in README to point to the private-source LBG1_LG3_I2_V1 environment
 
 ### Removed